Suppressed some conversion warnings with C++ style casts/type changes. JIRA CXX-8578
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@76239 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/services/netschedule_api_submitter.cpp
@@ -223,7 +223,7 @@ void CNetScheduleSubmitter::SubmitJobBatch(vector<CNetScheduleJob>& jobs,
                         "Invalid server response. Batch answer format.");
             }
 
-            port = atoi(s);
+            port = static_cast<short>(atoi(s));
             if (port == 0) {
                 NCBI_THROW(CNetServiceException, eProtocolError,
                         "Invalid server response. Port=0.");
